terraform/requirements.txt
boto3
python-dateutil
awscli